Adjust text area size, introduce character limit and add message components
Made text area in message-field component to have a fixed height and hidden overflow, to prevent it from growing infinitely. The character limit of textarea was set to 200 to stop extremely long messages. The color for text was tweaked to improve visibility. Also, a new index.tsx file was created inside the message component, which includes IncomingMessageComponent and OutgoingMessageComponent. These will be used for displaying both incoming and outgoing chat messages on the UI. Furthermore, the import statements for new components were added into src/pages/index.tsx file to be ready for future usage in pages.