Refactor FAB element and add it to public-groups
Refactored the Floating Action Button (FAB) to use styled components for position styling. A FabWrapper is introduced to place the FAB at a fixed position at the bottom right of the screen. This enhances the UI/UX by ensuring the add action is easily accessible from any part of the screen.

The refactored FAB was imported and used on the public-groups page, providing users the ability to add new groups easily while browsing existing ones. This encourages user interaction and engagement with the public groups feature.