Rename 'DialogConfirmDelete' component and refactor 'SelectOptions'
The `DialogConfirmDelete` component has been renamed to a more general `DialogConfirmAction`. It allows the cleaner reuse of the same UI element for confirming different types of operations, not only deletions.

`SelectOptions` component has been refactored to accept a `defaultOption` prop. This allows setting a selected option when the component gets rendered on the page. This change particularly improves usability in the context of an update form where stored values need to be pre-loaded.

Minor updates have been also applied to variable naming for consistency across the codebase.

Additionally, a new API mutation and the corresponding page for the update group operation have been created and implemented in the app. This allows users to edit group information, like group description or group type (private/public), or choose if the group should be deleted.