Refactor events page UI and enhance navigation
Refactored the events' page interface for better user experience and enhanced navigation inside the groups. On the events page, breadcrumbs navigation has been moved and is always shown on top. The display format of event dates has been localized. Changes also include renaming the label for groups based on whether they’re private or public. The CardImage component also displays the exact count of event pairs, instead of assuming it always exists. The alignment of the components was adjusted for better visual representation.