Added createEventApplication and eventApplicationPair queries
Introduced new queries for 'createEventApplication' and 'eventApplicationPair' in the GraphQL schema. Changes were necessary to support the creation of an event application and fetching a specific event application pair in the frontend. Also fixed some endpoints in routes, now using groupId rather than group name as parameter in "events?groupId". Scalar types for 'accountId', 'eventId', and 'applicationId' are corrected from 'Float' to 'Int'.