Add tabbed application view and error handling for event page
Event.tsx updated to show tabs for 'My Application' and 'Secret Santa Application' if there is an existing application for the event. GraphQL Query in graphql.tsx updated to retrieve application information for an event for the logged in user. Also enhanced error handling to accommodate additional connection error for the new query. This allows users to have a quick look at their applications directly from the event page.