Add account details to GraphQL queries and enforce auth check
Introduced the accountId to groups in graphql queries and added the avatarUrl to login queries. Also created a new getAccountCount query to fetch the total number of accounts. These changes enhance the user experience by providing more relevant information. In addition to this, an authentication check has been added to pages to ensure appropriate access control. Subsequently, reflected these changes in the corresponding schema and TypeScript definitions. This effort is intended to enhance both user experience and application security.