Implement gift received/not received functions and visual updates
Introduced new functionality to event.tsx to handle states indicating whether a gift was received or not. This provides interactive capabilities and feedback for the user. Two new dialog windows were introduced for the receipt of gift and if the gift was not received, providing clear user feedback and instructions. The application preference descriptions were also translated from English to Russian for better user experience for Russian-speaking audiences. Also, a minor change was provided in app-bar/index.tsx to visualize the avatar background as white. This enhances the look of the avatar image, providing a cleaner and more visually appealing user interface.