Add image upload and selection feature
Implemented a feature that allows user to upload image for group creation or update section as well for events creation. Image data is stored in `imageUrl` and handled by `handleImageChange` function. `axios` is used to make a POST request to the api and returns a link for the image. Errors during upload are caught and displayed. This allows the user to visualize what image they are uploading and makes sure the feature is user-friendly.