"Implement setEventApplicationStatus mutation
Added a new GraphQL mutation 'setEventApplicationStatus' which allows to update the status of an event application. This mutation was integrated into 'event.tsx', where it is used in 'onConfirmClick' handlers of two dialogs. This feature was necessary to handle user interactions for confirming that a gift was received or not received."