"Fix input email dialog and notifications functionality bugs
Corrected the 'handleSave' method in 'dialog-input-email' component to correctly pass the 'inputValue' to the parent component. Now, the user's input email is stored and passed, improving the overall UX, rather than a hardcoded 'email' string.
Also, fixed a typo in settings that incorrectly called the 'useEnableNotificationsMutation' hook in place of 'useDisableNotificationsMutation'. This corrects the issue of disabling notifications not functioning as expected.
Updated the showing of current email conditional to 'notifEnabled' and 'email', as 'email' could exist even when notifications are not enabled."