Add notifications feature and update GraphQL type
Implement a new feature for notifications that includes a new GraphQL query and a page to render these notifications. To adapt for the changes, we also adjust the types in our generated GraphQL types file.

Simultaneously, we remove an unnecessary check on our settings page that involves deleted accounts, as those users shouldn't be able to access settings to begin with.