Add dark mode functionality to application
In this commit, dark mode functionality was added to the TypeScript application. This was accomplished by adding a new state variable, 'darkMode', to track if dark mode should be enabled or not, and a 'toggleDarkMode' function to change the state. The theme is then selected based on the 'darkMode' state within the 'MaterialUiProvider' component. Also, a new button was added to the return of the component to allow users to toggle dark mode on and off. A new 'themeMuiDark' was also defined in the theme file to provide the necessary styles when dark mode is enabled.