Correct spelling mistakes and simple refactoring
Fixed typos in the localization files for both English and Russian languages. Particularly, the property 'untill_ending' has been corrected to 'until_ending'. Also, altered the text content in the application.json file to make them more grammatically correct and user-friendly. On the code side, removed unnecessary retrieval of 'event' in the 'createEventApplication' GraphQL mutation and added a 'justify-content' property in the 'DialogContainer's style to better center the dialog's buttons. The changes were made for overall enhancement of code quality and user experience.