Update chat UI and enable real-time updates
Made several key updates to the UI for chat and notifications, and enhanced real-time chat updates using socket.io. The Message component is now fixed at the bottom of the screen. For better layout management, a ContentContainer wraps components in the Page component.

In live chat, the app now implements locale-corrected date formatting, and new message alerts are managed by the Snackbar component in _app.tsx, providing a smoother user experience. Also, the layout for messages is revamped for enhanced readability. In Notifications, notifications are displayed in a more organized fashion for better user navigation.